+++ tests/rkt_non_root_test.go
@@ -104,7 +104,7 @@ func TestNonRootFetchRmGCImage(t *testing.T) {
 
 	rootImg := patchTestACI("rkt-inspect-root-rm.aci", "--exec=/inspect --print-msg=foobar")
 	defer os.Remove(rootImg)
-	rootImgHash := importImageAndFetchHash(t, ctx, rootImg)
+	rootImgHash := importImageAndFetchHash(t, ctx, "", rootImg)
 
 	// Launch/gc a pod so we can test non-root image gc.
 	runCmd := fmt.Sprintf("%s --insecure-options=image run --mds-register=false %s", ctx.Cmd(), rootImg)
@@ -125,7 +125,7 @@ func TestNonRootFetchRmGCImage(t *testing.T) {
 	// Should be able to remove the image fetched by ourselves.
 	nonrootImg := patchTestACI("rkt-inspect-non-root-rm.aci", "--exec=/inspect")
 	defer os.Remove(nonrootImg)
-	nonrootImgHash := importImageAndFetchHashAsGid(t, ctx, nonrootImg, gid)
+	nonrootImgHash := importImageAndFetchHashAsGid(t, ctx, nonrootImg, "", gid)
 
 	imgRmCmd = fmt.Sprintf("%s image rm %s", ctx.Cmd(), nonrootImgHash)
 	t.Logf("Running %s", imgRmCmd)