+++ tests/rkt_os_arch_test.go
@@ -161,7 +161,7 @@ func TestMissingOrInvalidOSArchFetchRun(t *testing.T) {
 	defer osArchTestRemoveImages(tests)
 
 	for i, tt := range tests {
-		imgHash := importImageAndFetchHash(t, ctx, tt.image)
+		imgHash := importImageAndFetchHash(t, ctx, "", tt.image)
 		rktCmd := fmt.Sprintf("%s run --mds-register=false %s", ctx.Cmd(), imgHash)
 		t.Logf("Running test #%v: %v", i, rktCmd)
 		runRktAndCheckOutput(t, rktCmd, tt.expectedLine, tt.expectError)